requirements.txt
pandas
duckdb
streamlit
sentence-transformers
faiss-cpu
dbt-core
dbt-duckdb
python-dotenv
numpy